Add files via upload
An interactive treemap was developed to examine the connections between census and survey data across various regions. The average of survey data was chosen as the box's color, while the average of census data served as the box's size, with region serving as the detail. The resulting map displayed the biggest and darkest blue box, which defined North America and represented the greatest average of both census and survey data. The Middle East and North Africa, as well as Sub-Saharan Africa, had light blue boxes that were similarly small, suggesting that their average census and survey statistics were lower than those of the other regions. The color encoding helped to emphasize disparities in the average census and survey data, and the treemap provided a useful representation for illustrating the relative sizes of various areas. The results indicate that while North America has the greatest levels of census and survey data, efforts to gather data in the Middle East and North Africa and Sub-Saharan Africa areas need to be improved if progress toward sustainable development is to be more accurately tracked.